Split runtime/mod.rs into dispatch, up, and state submodules
runtime/mod.rs keeps verb dispatch, stack-repo resolution, compose file
discovery, and mode detection; up.rs owns both up paths, port
allocation, the KNIT_* env contract, and database resolution; state.rs
owns run state, label-based down/status, and compose ps parsing.
Pure moves, no behavior change.

Knit-Group: kg_20260611_1f17ac
Knit-Bundle: open-source-prep
